Remove dashboard generation script.
Replace it with the exported (JSON) of the two primary dashboards we'll
be using: QmlBench Hardware & QmlBench Branches.

Branches dashboard is for use in comparing branch performance on a
single hardware target.

Hardware dashboard is for use in comparing a single branch across
multiple hardware targets.

Both of these dashboards are templated, and as such, no longer require
any static generation.