Annotations: Add support for adding & filtering on branch info to UI
Version updates and things like that only make sense for the branch they
happened on. So we don't want to see 5.6 events when looking at dev (for
example).

To fix this, we add a --branch parameter to annotate.py that takes a
string branch field. In the dashboards, we filter on this (allowing an
empty branch to indicate a "global" event), and we mark it as an
additional tag for UI purposes.