run.py: Add additional debug and fix handling of "bad" runs
Instead of skipping an empty run completely, report a 0 point. In my
opinion, this is better as it will give a visual report that something
is not right.

While we are here, use the reported CoV rather than generating it again
in the python runner to ensure there is only one source of values to
avoid future confusion.