Slight tweaks to visualization
Set the CoV to have a line fill, not the regular series.

With a lot of branches visible, the fills start to mess up the dashboard
a bit. Since the CoV are (ideally) at the bottom of the graph anyway,
having them filled is acceptable -- and it will provide a slight visual
indication if a CoV goes above the (10%) limit, above the graph.